+++ src/main/java/org/oogp/spring/CustomBeanFactory.java
@@ -1,10 +1,11 @@
-package org.oogp;
+package org.oogp.spring;
 
 import java.beans.PropertyEditor;
 import java.util.Set;
 import java.util.concurrent.Executor;
 import java.util.function.Consumer;
 
+import org.oogp.Classes;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.PropertyEditorRegistrar;
 import org.springframework.beans.PropertyEditorRegistry;
@@ -25,6 +26,7 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.core.ResolvableType;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.core.metrics.ApplicationStartup;
+import org.springframework.lang.Nullable;
 import org.springframework.util.StringValueResolver;
 
 /**
@@ -147,11 +149,13 @@ public class CustomBeanFactory implements AutowireCapableBeanFactory, Configurab
 	}
 
 	@Override
+	@Nullable
 	public Class<?> getType(final String name) throws NoSuchBeanDefinitionException {
 		return context.getType(name);
 	}
 
 	@Override
+	@Nullable
 	public Class<?> getType(final String name, final boolean allowFactoryBeanInit) throws NoSuchBeanDefinitionException {
 		return context.getType(name);
 	}
@@ -326,7 +330,7 @@ public class CustomBeanFactory implements AutowireCapableBeanFactory, Configurab
 	@Override
 	public BeanExpressionResolver getBeanExpressionResolver() {
 		// TODO: see if we need it
-		return (value, expressionContext) -> value;
+		return (value, _) -> value;
 	}
 
 	@Override