+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ---
 
+`1.0.2`
+
+- Added servers configuration.
+- Upgraded to Java 25.
+- Upgraded `jacoco-maven-plugin` to `0.8.14`.
+- Upgraded `apiphany` to `1.0.16`.
+- Upgraded `maven-plugin-plugin` to `3.15.2`.
+- Upgraded `maven-compiler-plugin` to `3.14.1`.
+- Added `schemaForObjectClass` to the plugin configuration which defaults to `object` not `string` as before.
+
+---
+
 `1.0.1`
 
 - Added `JavaEnvironment` to get Java environment information.